+++ Cargo.toml
@@ -4,6 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "oxi"
+path = "src/lib/mod.rs"
 
 [dependencies]
 assert_matches = "1.5.0"
