+++ src/lib/mod.rs
@@ -1,0 +1,3 @@
+pub mod settings;
+pub mod config;
+
